backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.4
PyJWT==2.8.0
redis==4.6.0
pytest==7.4.0
gunicorn==21.2.0

